Flytt hjelpefunksjoner fra typer til utils
